+++ go/src/internal/features/lab1/service/service.go
@@ -22,7 +22,7 @@ const (
 
 type StudentAttendanceItem struct {
 	StudentID  int
-	Attendance int
+	Attendance float64
 }
 
 type Student struct {
